+++ Overview/Settings/Components/WiggleModifier.swift
@@ -1,0 +1,54 @@
+/*
+ Settings/Components/WiggleModifier.swift
+ Overview
+ 
+ Created by William Pierce on 1/13/25.
+ 
+ Provides an attention-getting animation for warning indicators.
+*/
+
+import SwiftUI
+
+struct WiggleModifier: ViewModifier {
+    @State private var angle: Double = 0
+
+    func body(content: Content) -> some View {
+        content
+            .rotationEffect(.degrees(angle))
+            .onAppear {
+                withAnimation(.easeInOut(duration: 0.1)) {
+                    angle = 10
+                }
+
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.1) {
+                    withAnimation(.easeInOut(duration: 0.1)) {
+                        angle = -10
+                    }
+                }
+
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.2) {
+                    withAnimation(.easeInOut(duration: 0.1)) {
+                        angle = 10
+                    }
+                }
+
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
+                    withAnimation(.easeInOut(duration: 0.1)) {
+                        angle = -10
+                    }
+                }
+
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.4) {
+                    withAnimation(.easeInOut(duration: 0.1)) {
+                        angle = 10
+                    }
+                }
+
+                DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) {
+                    withAnimation(.easeInOut(duration: 0.1)) {
+                        angle = 0
+                    }
+                }
+            }
+    }
+}
